Added support for new Wishlist API
